@@ -1,14 +1,29 @@
-import Discord = require('discord.js');
+import { Client, GatewayIntentBits, ClientOptions, RESTPostAPIChatInputApplicationCommandsJSONBody, Collection, PermissionsBitField } from 'discord.js'
+
+import { } from 'discord.js';
 import { REST } from '@discordjs/rest';
 import { getVoiceConnection } from '@discordjs/voice';
-import { Routes } from 'discord-api-types/v9';
-import type { SlashCommandDataJSON } from '../typings/index';
-import { logger } from './log';
-import { getCommands, queryConfig } from './util';
+import { Routes } from 'discord-api-types/v10';
+import { queryConfig } from './util.js';
+import * as winston from 'winston';
+import { logger } from './log.js';
+import type SlashCommand from './commands/Command.js';
 
+class ClientExtend extends Client<true, any> {
+    public commands: Collection<string, SlashCommand>;
+    public logger: winston.Logger;
+
+    constructor(options: ClientOptions) {
+        super(options);
+        this.logger = logger;
+        this.commands = new Collection();
+    }
+}
+
+// todo: initialize client with other options
 const rest = new REST({ version: '9' });
-const client = new Discord.Client({
-    intents: ['GUILDS', 'GUILD_MESSAGES', 'GUILD_VOICE_STATES', 'GUILD_MESSAGE_REACTIONS', 'GUILD_MEMBERS']
+const client = new ClientExtend({
+    intents: [GatewayIntentBits.Guilds, GatewayIntentBits.GuildMessages, GatewayIntentBits.GuildVoiceStates, GatewayIntentBits.GuildMessageReactions, GatewayIntentBits.GuildMembers]
 });
 
 enum ClientEnums {
@@ -18,11 +33,6 @@ enum ClientEnums {
     TEST_CLIENT_ID = '464186918457049088',
     PROD_CLIENT_ID = '783886978974220338',
 }
-
-client.commands = new Map();
-client.musicQueueManager = new Discord.Collection;
-client.audioPlayers = new Map();
-client.logger = logger;
 
 /** Retrieve token from Firestore */
 queryConfig().then(config => {
@@ -46,22 +56,24 @@ client.once('ready', async () => {
             message: `Succesfully logged into ${client.user.tag}`
         });
 
-    const commands = await getCommands();
-    client.commands = commands;
+    // import commands dynamically
+    // todo: multiple commands file handler
+    const ExampleCommand = await import(''+'./commands/test.js');
+    const cmd: SlashCommand = ExampleCommand.default;
 
-    let data: SlashCommandDataJSON[] = [];
-    for (const [k, v] of commands) {
-        data.push(v.data.toJSON())
-    }
+    let data: RESTPostAPIChatInputApplicationCommandsJSONBody[] = [];
+
+    data.push(cmd.data.toJSON());
+    client.commands.set(cmd.data.name, cmd);
 
     if (process.env.NODE_ENV == 'dev') {
-        /** debugging guilds */
-        await rest.put(Routes.applicationGuildCommands(ClientEnums.TEST_CLIENT_ID, ClientEnums.TANK_SERVER_ID), { body: data });
+        // debugging guilds
+        //await rest.put(Routes.applicationGuildCommands(ClientEnums.TEST_CLIENT_ID, ClientEnums.TANK_SERVER_ID), { body: data });
         await rest.put(Routes.applicationGuildCommands(ClientEnums.TEST_CLIENT_ID, ClientEnums.DEV_SERVER_ID), { body: data });
     } else {
         await rest.put(Routes.applicationCommands(ClientEnums.PROD_CLIENT_ID), { body: data });
     }
-
+    
     await client.application?.commands.fetch();
     client.logger.log({
         level: 'info',
@@ -73,29 +85,40 @@ client.once('ready', async () => {
 /** Main command handler */
 client.on('interactionCreate', async interaction => {
     if (!interaction.isCommand() || !interaction.guild) return;
+    
+    const interactionCommand = interaction.command;
     const command = interaction.commandName;
+
     try {
-        client.commands.get(command)!.execute(interaction);
-        client.logger.log({
-            level: 'verbose',
-            label: 'main',
-            guild: interaction.guild,
-            channel: interaction.channel,
-            user: interaction.user,
-            message: `Ran command ${command}`
-        });
+        
+        /* Check if command actually exists and has a function */
+        if (!(client.commands.get(command)) ||
+            (client.commands.get(command)?.execute == undefined)) {
+                console.log(client.commands.get(command));
+                throw `Command ${command} does not exist in client`;
+        } else {
+            client.commands.get(command)!.execute(interaction);
+            client.logger.log({
+                level: 'verbose',
+                label: 'main',
+                guild: interaction.guild,
+                channel: interaction.channel,
+                user: interaction.user,
+                message: `Ran command ${command}`
+            });
+        }
     } catch (e) {
         client.logger.log({
             level: 'error',
             label: 'main',
-            message: e.stack
+            message: e.stack ?? e
         });
     }
 });
 
 /** Remove all commands and kill process */
 client.on('messageCreate', async message => {
-    const hasPermissions = message.member?.permissions.has(Discord.Permissions.FLAGS.ADMINISTRATOR);
+    const hasPermissions = message.member?.permissions.has(PermissionsBitField.Flags.Administrator);
     if (hasPermissions && message.content.startsWith('.refresh')) {
         client.logger.log({
             level: 'info',
@@ -106,36 +129,18 @@ client.on('messageCreate', async message => {
             for (const cmd of cmds) await client.application?.commands.delete(cmd[1]);
         }).finally(() => process.exit(-1));
     }
-
-    // quote channel cleaning up
-    if (message.channel.id == '747290626038431875' && !message.author.bot) {
-        if (!message.content.match(/^['">]/gi)) {
-            let response = await message.reply({ content: `\`${message.content}\` doesn't appear to be quoting anything. Perhaps use a reaction instead?\nThis message will delete itself in 10 seconds..`});
-            setTimeout(() => {
-                response.delete();
-                if (message && message.deletable) message.delete();
-            }, 10000);
-        }
-    }
-
-    if (message.channel.id == '951924418408505434') {
-        // delete message after 1 minute
-        setTimeout(() => { message.delete(); }, 60 * 1000);
-    }
 });
 
-/** AFK Timeout (5 minutes) */
+/** VC AFK Timeout (5 minutes) */
 client.on('voiceStateUpdate', (pre) => {
-    let connection = getVoiceConnection(pre.guild.id),
-        queue = client.musicQueueManager.get(pre.guild.id)!;
-    if (!queue) return;
-    else if ((queue.length == 0) && connection) {
+    let connection = getVoiceConnection(pre.guild.id);
+    if (connection) {
         const channelId = connection!.joinConfig.channelId,
             beforeChannelId = pre.channel?.id,
             isAlone = (!(pre.channel?.members.size) || pre.channel.members.size < 2);
         if (channelId == beforeChannelId && isAlone) {
             /** Leave channel after 5 minutes */
-            queue.leaveTimeout = setTimeout(() => { connection!.destroy() }, ClientEnums.AFK_TIMEOUT_MINUTES * 60 * 1000); 
+            setTimeout(() => { connection!.destroy() }, ClientEnums.AFK_TIMEOUT_MINUTES * 60 * 1000); 
         }
     }
 });
