@@ -8,80 +8,22 @@ import type { MusicQueue } from '../src/commands/utils/musicUtils';
 
 declare module 'discord.js' {
     interface Client<T> {
-        commands: Map<string, DiscordCommand>,
-        audioPlayers: Map<string, AudioPlayerWithResource>
-        musicQueueManager: Collection<string, MusicQueue>,
         cloudProjectId?: string,
         logger: Logger
     };
-    /** for our purposes, guildId will always be provided */
+
     interface CommandInteraction {
         guildId: string
         channel: TextBasedChannel
     }
 }
 
-interface AudioPlayerWithResource {
-    player: AudioPlayer
-    resource: AudioResource
-}
 interface BotConfig {
     token: string,
     testToken: string,
 }
 
-type VideoResult = Omit<ytsr.Video, 'bestThumbnail' | 'author' | 'isUpcoming' | 'views' |'isLive' | 'badges' | 'isComing' | 'upcoming' | 'uploadedAt' | 'description'> 
-
 interface embed {
     name: string,
     createEmbed(...args): MessageEmbed;
-}
-
-interface SlashCommandDataJSON {
-    name: string;
-    description: string;
-    options: APIApplicationCommandOption[];
-    default_permission: boolean | undefined;
-}
-interface DiscordCommand extends ApplicationCommand {
-    data: SlashCommandBuilder
-    toJSON(): SlashCommandDataJSON
-    category: 'Admin' | 'General' | 'Music' | 'Experimental',
-    execute(...args): any
-}
-
-interface MinecraftResponse {
-    description: {
-        text: string
-    },
-    players: {
-        max: number,
-        online: number,
-        sample: [ name: string, id: string ]
-    }
-    version: {
-        name: string,
-        protocol: number,
-    }
-    favicon?: string,
-    ping: number,
-}
-
-type BotCommands = Collection<string, DiscordCommand>;
-
-interface geminiResponse {
-    symbol: string,
-    open: string,
-    high: string,
-    low: string,
-    close: string,
-    changes: string[],
-    bid: string,
-    ask: string,
-}
-
-interface cryptoInfo extends geminiResponse {
-    exchange: string,
-    fiat: string,
-    ticker: string,
 }
